requirements.txt
langgraph>=0.0.26
langchain-core>=0.1.30
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
langchain-openai>=0.1.0
python-dotenv>=1.0.0
